chat/requirements.txt
django
python-dotenv
requests
gunicorn
pyngrok
djangorestframework
psycopg2-binary
